Restructure project
- Use a single main.c with the logic
- Use Makefile macro flag to optionally include the custom library
- Rename the custom string.h function with my_ prefix
- Add my_string_names.h for function names alias
- Move supplemental functions from main.c to functions.c
- Update Makefile accordingly
- Remove old files no longer used

+++ my_string.h
@@ -6,7 +6,7 @@
 #ifndef MY_STRING_H
 #define MY_STRING_H
 
-int my_strlen(const char*);
+size_t my_strlen(const char*);
 
 
 #endif
